+++ tests/test_br.py
@@ -3,6 +3,7 @@ from __future__ import unicode_literals
 
 from django.test import SimpleTestCase
 
+from localflavor.br import models
 from localflavor.br.forms import (BRCNPJField, BRCPFField, BRProcessoField, BRStateChoiceField, BRStateSelect,
                                   BRZipCodeField)
 
@@ -177,3 +178,46 @@ class BRLocalFlavorTests(SimpleTestCase):
             'pr': error_invalid,
         }
         self.assertFieldOutput(BRStateChoiceField, valid, invalid)
+
+
+class BRLocalFlavorModelTests(SimpleTestCase):
+
+    def test_BRCNPJField(self):
+        instance = models.BRCNPJField()
+        name, path, args, kwargs = instance.deconstruct()
+        new_instance = models.BRCNPJField(*args, **kwargs)
+        self.assertEqual(instance.max_length, new_instance.max_length)
+        self.assertEqual(instance.description, new_instance.description)
+        self.assertEqual(instance.validators, new_instance.validators)
+
+    def test_BRCPFField(self):
+        instance = models.BRCPFField()
+        name, path, args, kwargs = instance.deconstruct()
+        new_instance = models.BRCPFField(*args, **kwargs)
+        self.assertEqual(instance.max_length, new_instance.max_length)
+        self.assertEqual(instance.description, new_instance.description)
+        self.assertEqual(instance.validators, new_instance.validators)
+
+    def test_BRCellPhoneField(self):
+        instance = models.BRCellPhoneField()
+        name, path, args, kwargs = instance.deconstruct()
+        new_instance = models.BRCellPhoneField(*args, **kwargs)
+        self.assertEqual(instance.max_length, new_instance.max_length)
+        self.assertEqual(instance.description, new_instance.description)
+        self.assertEqual(instance.validators, new_instance.validators)
+
+    def test_BRTelephoneField(self):
+        instance = models.BRTelephoneField()
+        name, path, args, kwargs = instance.deconstruct()
+        new_instance = models.BRTelephoneField(*args, **kwargs)
+        self.assertEqual(instance.max_length, new_instance.max_length)
+        self.assertEqual(instance.description, new_instance.description)
+        self.assertEqual(instance.validators, new_instance.validators)
+
+    def test_BRZipCodeField(self):
+        instance = models.BRZipCodeField()
+        name, path, args, kwargs = instance.deconstruct()
+        new_instance = models.BRZipCodeField(*args, **kwargs)
+        self.assertEqual(instance.max_length, new_instance.max_length)
+        self.assertEqual(instance.description, new_instance.description)
+        self.assertEqual(instance.validators, new_instance.validators)